# api.yaml
openapi: 3.0.0
info:
  title: endeus recipe API
  version: 1.0.0
  contact:
    name: Victor Santoso
    email: victorsantoso1306@gmail.com
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/register:
    post:
      summary: Register a new user
      description: Register a new user with the provided information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/requestBodies/PostRegisterRequestBody'
            example:
              role: ADMIN
              email: testguser@gmail.com
              password: testUser*999
              name: Test User
      responses:
        '200':
          description: Success response for Register Endpoint
          content:
            application/json:
              schema:
                $ref: '#/components/responses/PostRegisterSuccessResponse'
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
                code: 200
                message: successfully registered a new user
        '400':
          description: Bad Request response error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
              example:
                message: bad request
                code: 400
        '409':
          description: Conflict response error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
              example:
                message: duplicate entry
                code: 409
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
              example:
                message: internal server error
                code: 500
  /api/v1/login:
    post:
      summary: Log in a user
      description: Log in a user with the provided credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/requestBodies/PostLoginRequestBody'
            example:
              email: testuser@gmail.com
              password: testUser*999
      responses:
        '200':
          description: Success response for Log in Endpoint
          content:
            application/json:
              schema:
                $ref: '#/components/responses/PostLoginSuccessResponse'
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
                code: 200
                message: successfully logged in
        '400':
          description: Bad Request response error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
              example:
                message: bad request
                code: 400
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
              example:
                message: internal server error
                code: 500
# RECIPE 
...
components:
  requestBodies:
    PostRegisterRequestBody:
      description: Request body for user registration endpoint.
      content:
        application/json:
          schema:
            type: object
            required:
              - role
              - email
              - password
              - name
            properties:
              role:
                type: string
                enum:
                  - ADMIN
                  - READER
                description: Role for defining user access permission.
              email:
                type: string
                format: email
                minLength: 9
                maxLength: 60
                description: User's email for registration.
              password:
                type: string
                minLength: 6
                maxLength: 20
                description: User's password for registration.
              name:
                type: string
                minLength: 3
                maxLength: 60
                description: User's full name for registration.
              profile_image:
                type: string
                description: User's profile image in url format (for demo purposes). In real case you might want to use something like cloud storage services to store profile_image.
    PostLoginRequestBody:
      description: Request body for login endpoint.
      content:
        application/json:
          schema:
            type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
                format: email
                description: User's email for login.
              password:
                type: string
                description: User's password for login.
...
  responses:
    PostRegisterSuccessResponse:
      description: Successful registration response.
      content:
        application/json:
          schema:
            type: object
            properties:
              access_token:
                type: string
                description: Access token for the registered user.
              message:
                type: string
                description: Message indicating the success of the registration process.
              code:
                type: integer
                format: int32
                description: Code indicating the success of the registration process.
    PostLoginSuccessResponse:
      description: Successful login response.
      content:
        application/json:
          schema:
            type: object
            properties:
              access_token:
                type: string
                description: Access token for the logged-in user.
              message:
                type: string
                description: Message indicating the success of the login process.
              code:
                type: integer
                format: int32
                description: Code indicating the successs of the login process.
...
    ErrorResponse:
      description: Response upon error process.
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                description: Error code for api response based on error.
              message:
                type: string
                description: Error message for api response based on error.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...